Won't tell other label index to drop if read-only

+++ community/kernel/src/main/java/org/neo4j/kernel/extension/dependency/DeleteStoresFromOtherLabelScanStoreProviders.java
@@ -53,7 +53,10 @@ public class DeleteStoresFromOtherLabelScanStoreProviders implements DependencyR
             {
                 try
                 {
-                    provider.drop();
+                    if ( !provider.isReadOnly() )
+                    {
+                        provider.drop();
+                    }
                 }
                 catch ( IOException e )
                 {
